+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.0
+
+### Fixed
+
+-   Dark mode.
+
+### Added
+
+-   Filter forms.
+-   Defer loading.
+-   Loading indicator.
+-   Widget height.
+-   Widget footer.
+-   Dark mode.
+
+### Changed
+
+-   Documentation.
+
 ## 0.1.5
 
 ### Fixed